Downgrade to musl 1.2.3.
Starting with musl 1.2.4 the *64 aliases (open64, etc.) are no longer
available (see commit 246f1c811448f37a44b41cd8df8d0ef9736d95f4). This is
a problem for us, because we are currently using target_env=gnu instead
of target_env=musl for Rust applications on Linux. Obviously, this is a
hack, but with target_env=musl it doesn't work at all with RISC-V. I
think this is simply not ready to use.