musl: upgrade to 1.2.5 and re-enable LSF64.
We need a recent musl version in order to get riscv32 support.
Unfortunately, musl dropped the LSF64 symbols (open64, etc.) which we
need for M³Linux. The issue is that RISCV+musl is not supported yet by
the Rust standard library. We therefore pretend to use the glibc, which
works interestingly. However, when doing so the LSF64 symbols are
required. We therefore re-enable these symbols by reverting the patch
that removed them.